Add two examples to the C API repo
PS logfile to pcap file type convertor tool. This is a great practicle example of the logfile iterator. The application converts a PS plog file to a pcap file capable of being played in Velodyne's Veloview application.

Logfile iterator for the Velodyne HDL, which shows how to cast to the appropriate data strcutres, and access the sensors raw data.