User data model Makefile changes for 2.0
Prior to this commit the user data model example did not reflect the
state of a polysync install under version 2.0. After this commit it
will. polysync data model libs now live in `/usr/lib` and need to
be installed there with `sudo`. Installing the rest of the artifacts to
`/usr/local/polysync` also now requires `sudo`. The CPP api is no
longer generated by default, a `-c` flag must be passed to pdm-gen
to enable it and that flag was added to the generation step in
the Makefile.