Logfile itertor for video device plogs
This example allows a user to iterate over a video-device plog. If the
logged pixel format is YUYV this example provides functionality to
generate image files (bmp or ppm) for every raw video frame.

See README.md for instructions on building, dependencies and usage.

Further usage information can be accessed with a `-?` argument to the
executable.